Start converting editor.js
